BPE26-ref: extract feature injection function to dedicated file
