+++ src/content.ts
@@ -1,26 +1,16 @@
 // Plume - TypeScript for song page and album page display
 import { BC_ELEM_IDENTIFIERS, DebugControl } from "./domain/bandcamp";
-import { APP_NAME, APP_VERSION, PLUME_KO_FI_URL } from "./domain/meta";
 import { PLUME_ELEM_IDENTIFIERS } from "./domain/plume";
 import { setupAudioEventListeners } from "./features/audio-events";
 import { cleanupFullscreenMode, toggleFullscreenMode } from "./features/fullscreen";
 import { getString, logDetectedBrowser } from "./features/i18n";
+import { injectEnhancements } from "./features/injection";
 import { setupHotkeys } from "./features/keyboard";
 import { CPL, logger } from "./features/logger";
-import {
-  findOriginalPlayerContainer,
-  hideOriginalPlayerElements,
-  restoreOriginalPlayerElements,
-} from "./features/original-player";
-import { getInfoSectionWithRuntime } from "./features/runtime";
 import { setupStoreSubscriptions } from "./features/store-subscriptions";
 import { getTrackQuantifiers } from "./features/track-quantifiers";
-import { getAppropriatePretextColor, getCurrentTrackTitle } from "./features/track-title";
+import { getCurrentTrackTitle } from "./features/track-title";
 import {
-  createFullscreenButtonSection,
-  createPlaybackControlPanel,
-  createProgressBar,
-  createVolumeControlSection,
   handlePlayPause,
   handleTimeBackward,
   handleTimeForward,
@@ -30,7 +20,6 @@ import {
 } from "./features/ui";
 import { getPlumeUiInstance, PLUME_ACTION_TYPES } from "./infra/AppInstanceImpl";
 import { getStoreInstance, STORE_ACTION_TYPES } from "./infra/AppStoreImpl";
-import { PLUME_SVG } from "./svg/icons";
 
 (() => {
   "use strict";
@@ -63,11 +52,6 @@ import { PLUME_SVG } from "./svg/icons";
     logger(CPL.INFO, `${getString("INFO__VOLUME__FOUND")} ${Math.round(volume * 100)}${getString("META__PERCENTAGE")}`);
 
     return audio;
-  };
-
-  const addRuntime = () => {
-    const trackView = document.querySelector(BC_ELEM_IDENTIFIERS.trackView) as HTMLDivElement;
-    trackView.insertBefore(getInfoSectionWithRuntime(), trackView.firstChild);
   };
 
   // Debug function to identify Bandcamp controls
@@ -128,92 +112,6 @@ import { PLUME_SVG } from "./svg/icons";
     if (!currentTrackTitleElem) return false;
 
     return lastTrackTitleElem?.textContent === currentTrackTitleElem.textContent;
-  };
-
-  const injectEnhancements = async () => {
-    const bcPlayerContainer = findOriginalPlayerContainer();
-    if (!bcPlayerContainer) {
-      logger(CPL.ERROR, getString("ERROR__UNABLE_TO_FIND_CONTAINER"));
-      return;
-    }
-
-    const isAlbumPage = store.getState().pageType === "album";
-
-    restoreOriginalPlayerElements(); // call it to prevent "unused function" linter warning
-    hideOriginalPlayerElements();
-
-    const plumeContainer = document.createElement("div");
-    plumeContainer.id = PLUME_ELEM_IDENTIFIERS.plumeContainer.split("#")[1];
-
-    const headerContainer = document.createElement("div");
-    headerContainer.id = PLUME_ELEM_IDENTIFIERS.headerContainer.split("#")[1];
-
-    const headerLogo = document.createElement("a");
-    headerLogo.id = PLUME_ELEM_IDENTIFIERS.headerLogo.split("#")[1];
-    headerLogo.innerHTML = PLUME_SVG.logo + `<p id="${headerLogo.id}__version">${APP_VERSION}</p>`;
-    headerLogo.href = PLUME_KO_FI_URL;
-    headerLogo.target = "_blank";
-    headerLogo.rel = "noopener noreferrer";
-    headerLogo.ariaLabel = APP_NAME;
-    headerLogo.title = getString("ARIA__LOGO_LINK");
-    headerContainer.appendChild(headerLogo);
-
-    const initialTrackTitle = getCurrentTrackTitle(isAlbumPage);
-    const initialTq = getTrackQuantifiers(initialTrackTitle);
-    const currentTitleSection = document.createElement("div");
-    currentTitleSection.id = PLUME_ELEM_IDENTIFIERS.headerCurrent.split("#")[1];
-    currentTitleSection.tabIndex = 0; // make it focusable for screen readers
-    currentTitleSection.ariaLabel = isAlbumPage
-      ? getString("ARIA__TRACK_CURRENT", [initialTq.current, initialTq.total, initialTrackTitle])
-      : getString("ARIA__TRACK", initialTrackTitle);
-    const currentTitlePretext = document.createElement("span");
-    currentTitlePretext.id = PLUME_ELEM_IDENTIFIERS.headerTitlePretext.split("#")[1];
-    const initialTrackNumberText = isAlbumPage
-      ? getString("LABEL__TRACK_CURRENT", `${initialTq.current}/${initialTq.total}`)
-      : getString("LABEL__TRACK");
-    currentTitlePretext.textContent = initialTrackNumberText;
-    currentTitlePretext.style.color = getAppropriatePretextColor();
-    currentTitlePretext.ariaHidden = "true"; // hide from screen readers to avoid redundancy
-    currentTitleSection.appendChild(currentTitlePretext);
-    const currentTitleText = document.createElement("span");
-    currentTitleText.id = PLUME_ELEM_IDENTIFIERS.headerTitle.split("#")[1];
-    currentTitleText.textContent = initialTrackTitle;
-    currentTitleText.title = initialTrackTitle; // see full title on hover in case title is truncated
-    currentTitleText.ariaHidden = "true"; // hide from screen readers to avoid redundancy
-    currentTitleSection.appendChild(currentTitleText);
-    headerContainer.appendChild(currentTitleSection);
-
-    // Initialize store with current track title and track number
-    store.dispatch({ type: STORE_ACTION_TYPES.SET_TRACK_TITLE, payload: initialTrackTitle });
-    store.dispatch({ type: STORE_ACTION_TYPES.SET_TRACK_NUMBER, payload: initialTrackNumberText });
-
-    plumeUiInstance.dispatch({ type: PLUME_ACTION_TYPES.SET_TITLE_DISPLAY, payload: headerContainer });
-    plumeContainer.appendChild(headerContainer);
-
-    const playbackManager = document.createElement("div");
-    playbackManager.id = PLUME_ELEM_IDENTIFIERS.playbackManager.split("#")[1];
-
-    const progressContainer = createProgressBar();
-    playbackManager.appendChild(progressContainer);
-    const playbackControls = createPlaybackControlPanel();
-    if (playbackControls) {
-      playbackManager.appendChild(playbackControls);
-    }
-    plumeContainer.appendChild(playbackManager);
-
-    const volumeContainer = await createVolumeControlSection();
-    if (volumeContainer) {
-      plumeContainer.appendChild(volumeContainer);
-    }
-
-    const fullscreenBtnSection = createFullscreenButtonSection(toggleFullscreenMode);
-    plumeContainer.appendChild(fullscreenBtnSection);
-
-    bcPlayerContainer.appendChild(plumeContainer);
-
-    logger(CPL.LOG, getString("LOG__MOUNT__COMPLETE"));
-
-    if (isAlbumPage) addRuntime();
   };
 
   const updateTrackForwardBtnState = () => {
